requirements/base.txt
# ── Base — diinstall oleh SEMUA service ───────────────────────────────────────
celery==5.4.0
redis==5.2.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
requests==2.32.3
aiohttp==3.10.10
psutil==6.1.0
tqdm==4.67.0

# Vector DB
qdrant-client==1.12.1

# Gemini (embedding + LLM fallback)
google-generativeai==0.8.3
google-genai==1.3.0

# Token counting
tiktoken==0.8.0